Separate state manipulation logic into Reducers module Closes #4
